Fix Visual Studio 2012 build errors.
Also, remove execute permissions on several source files.

Tested: Builds on Visual Studio 2012, and Linux.

+++ src/components/patron.cpp
@@ -60,7 +60,7 @@ using motive::MotiveTime;
 using motive::kMotiveTimeEndless;
 
 // All of these numbers were picked for purely aesthetic reasons:
-static const float kSplatterCount = 10;
+static const int kSplatterCount = 10;
 static const float kLapWaitAmount = 0.5f;
 static const float kMaxCatchTime = 1.5f;
 static const float kHeightRangeBuffer = 0.05f;
